Restart checks to see if service is running before attempting kill of pceleryd

+++ ptmscout_web/restart.sh
@@ -4,8 +4,11 @@ sudo chmod -R 775 *
 sudo chown -R www-data *
 sudo chgrp -R development *
 
-kill `cat "logs/pids"`
-/data/pyramid/bin/pceleryd development.ini --logfile=logs/pceleryd --loglevel=DEBUG --pidfile=logs/pids > /dev/null &
+if [ -e "logs/pids" ]
+then
+    kill `cat "logs/pids"`
+fi
 
+/data/pyramid/bin/pceleryd development.ini --logfile=logs/pceleryd --loglevel=DEBUG --pidfile=logs/pids > /dev/null &
 sudo apache2ctl restart
